Use shift+wasd/mouse for flying similar to Star Rod
This removes the right-mouse-down hidden cursor feature which should
probably be enabled whilst shift is held, but for the moment it's good.
These controls are how Star Rod handles it and we should probably use
those by default as it will be what most modders are used to.

Also the bitfield for keys thing didn't work on my Linux machine so
we're now using glium's abstraction, `VirtualKeyCode`. and a hashmap.